# use fnv instead of SipHasher for HashMaps. add type FnvMap