--- a/base/src/fnv.rs
+++ b/base/src/fnv.rs
@@ -1,0 +1,8 @@
+extern crate fnv;
+
+use std::collections::HashMap;
+use std::hash::BuildHasherDefault;
+
+pub use self::fnv::FnvHasher;
+
+pub type FnvMap<K, V> = HashMap<K, V, BuildHasherDefault<FnvHasher>>;