--- a/base/src/scoped_map.rs
+++ b/base/src/scoped_map.rs
@@ -4,7 +4,7 @@ use std::collections::hash_map;
 use std::collections::hash_map::{Entry, IterMut};
 use std::hash::Hash;
 
-use types::FnvMap;
+use fnv::FnvMap;
 
 /// A map struct which allows for the introduction of different scopes
 /// Introducing a new scope will make it possible to introduce additional