@brendanzab  mentioned putting the FnvMap alias elsewhere, the map appears to be most used in base/src though. I'm open to any suggestions. 

All tests pass.

Fixes #95