@leshow 2016-08-09
Sure, I'll move it out of `types` and add to the PR