@brendanzab 2016-08-09
Just so you understand my reasoning, the name `types` refers to the fact that these are the data structures used for gloun's typechecking.  The naming might be initially confusing though as many Rust libraries use the same name to contain their common data structures.