@leshow 2016-08-09
Yep, I follow, I appreciate your guidance, thanks!

I've run the tests again following the refactor, everything seems fine.