@brendanzab 2016-08-09
In your PR description, can you add 'Fixes #95'? That will mean that the issue will be auto-closed when we merge.

You also might be interested in joining our gitter channel (if you didn't know about it already) where we have more real-time discussions: https://gitter.im/Marwes/gluon