@Marwes 2016-08-09
> In your PR description, can you add 'Fixes #95' at the top?

:+1:  Otherwise it is also possible to add this in the merge commit.