@Marwes 2016-08-09
The results vary quite a lot but for typechecking I am seeing a 5-10% speedup. Probably a bit overshadowed by inneficiencies elsewhere but its still a nice speedup!

Before
```
test clone_prelude     ... bench:     248,752 ns/iter (+/- 142,182)
test typecheck_prelude ... bench:  12,221,382 ns/iter (+/- 3,601,499)
test prelude ... bench:  19,270,757 ns/iter (+/- 1,817,400)
```
After
```
test clone_prelude     ... bench:     252,140 ns/iter (+/- 159,846)
test typecheck_prelude ... bench:  11,597,207 ns/iter (+/- 2,093,485)
test prelude ... bench:  18,961,071 ns/iter (+/- 3,236,183)
```
(clone_prelude measure overhead due to cloning the AST every iteration in `typecheck_prelude`)