@Marwes 2016-08-09
I seem to have caused caused some merge conflicts by merging #102, sorry!